pyproject.toml
[tool.poetry]
name = "crewai-research-assistant"
version = "0.1.0"
description = "A CrewAI-based franchise research assistant"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "crewai_research_assistant", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
crewai = {extras = ["tools"], version = ">=0.76.2"}
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"
langchain = ">=0.0.350"
duckduckgo-search = ">=4.1.1"
google-search-results = ">=2.4.2"
beautifulsoup4 = ">=4.12.2"
playwright = ">=1.40.0"
markdown = ">=3.5.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
pytest-cov = "^4.1.0"
mypy = "^1.7.0"
readme-md-generator = "^1.0.0"
readme-scribe = "^0.0.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
research = "crewai_research_assistant.main:run"
train-crew = "crewai_research_assistant.main:train"
replay-crew = "crewai_research_assistant.main:replay"
test-crew = "crewai_research_assistant.main:test"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi-line-output = 3

[tool.mypy]
python-version = "3.10"
warn-return-any = true
warn-unused-configs = true
disallow-untyped-defs = true
check-untyped-defs = true

[tool.pytest.ini-options]
testpaths = ["tests"]
python-files = ["test_*.py"]
addopts = "-v --cov=src"
